requirements.txt
absl-py==2.1.0
accelerate==1.11.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.11.0
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.5
black==23.11.0
blis==1.2.1
brotli==1.2.0
build==1.3.0
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
coverage==7.10.7
crewai==1.4.1
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.5.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
doctr==1.9.0
durationpy==0.10
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
evaluate==0.4.6
executing==2.2.0
faiss-cpu==1.13.0
fastapi==0.121.2
ffmpy==1.0.0
filelock==3.20.0
flake8==7.3.0
flatbuffers==25.9.23
fonttools==4.55.5
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2025.10.0
gdown==5.2.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.72.0
gradio==5.49.1
gradio_client==1.13.3
greenlet==3.2.4
groovy==0.1.2
grpcio==1.70.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.0.0
instructor==1.13.0
ipykernel==6.29.5
ipython==8.31.0
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.4.2
json5==0.12.1
json_repair==0.25.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kubernetes==34.1.0
langchain==1.0.7
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.4
langchain-ollama==1.0.0
langchain-text-splitters==1.0.0
langcodes==3.5.0
langdetect==1.0.9
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.42
language_data==1.3.0
lightning-utilities==0.15.2
lxml==6.0.2
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.1.7
mccabe==0.7.0
mcp==1.21.1
mdurl==0.1.2
ml_dtypes==0.5.3
mmh3==5.2.0
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy==1.7.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.6.1
onnx==1.19.1
onnxruntime==1.23.2
openai==2.8.0
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
parso==0.8.4
pathspec==0.12.1
pdfminer.six==20251107
pdfplumber==0.11.8
peft==0.17.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.4.0
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==6.33.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.4.2
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydub==0.25.1
pyflakes==3.4.0
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==7.4.4
pytest-cov==7.0.0
pytest-order==1.2.0
python-dateutil==2.9.0.post0
python-doctr==1.0.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.3
pyzmq==26.2.0
RapidFuzz==3.14.3
referencing==0.37.0
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rouge_score==0.1.2
rpds-py==0.28.0
rsa==4.9
ruff==0.14.5
sacrebleu==2.5.1
safehttpx==0.1.7
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.16.3
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==75.8.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.5.1
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.49.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.15.1
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.5.0
tokenizers==0.22.1
tomli==2.3.0
tomli_w==1.2.0
tomlkit==0.13.3
torch==2.2.2
torchmetrics==1.8.2
torchvision==0.17.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
ty==0.0.1a26
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.5.0
uv==0.9.9
uvicorn==0.38.0
uvloop==0.22.1
validators==0.35.0
virtualenv==20.35.4
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.23.0
zstandard==0.25.0
